Remove .tagged(...) method (better dev flow)
Making a developer move between .with(...) and .tagged(...) if they add
or remove data to a component adds friction when it should be easy. Also
TypeScript is not able to show an error once the developer does add a
factory function to a previously .tagged(...) component, making this
even worse.